make MemberConverter static
